tests/wpt/web-platform-tests/css/css-writing-modes/parsing/direction-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Writing Modes: getComputedValue().direction</title>
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#propdef-direction">
<meta name="assert" content="direction computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("direction", "ltr");
test_computed_value("direction", "rtl");
</script>
</body>
</html>
